+++ ChangeLog.md
@@ -12,6 +12,10 @@
   ([#75](https://github.com/davep/hike/pull/75))
 - Fixed correct document not being shown when a command is given on the
   command line. ([#76](https://github.com/davep/hike/issues/76))
+- Added an unattractive but better-than-nothing workaround for [yet another
+  Textual `OptionList`
+  borkage](https://github.com/Textualize/textual/issues/5701).
+  ([#77](https://github.com/davep/hike/issues/77))
 
 ## v0.9.0